One translation framework al update to dev (#41)
* Update to dev commit 3cdea62c31b1be60aa02138e. Doesn't run yet, issues with install

* Fixed install for AL plugin

* Fixed broken translators and variant switching

+++ plugin/al/mayautils/AL/maya/event/MayaEventManager.h
@@ -15,10 +15,8 @@
 //
 #pragma once
 
-#include "./Api.h"
+#include "AL/maya/event/Api.h"
 
-#include <string>
-#include <vector>
 #include "maya/MCommandMessage.h"
 #include "maya/MDagMessage.h"
 #include "maya/MPaintMessage.h"
@@ -365,7 +363,7 @@ enum UiMessage
 class MayaEventHandler
   : public AL::event::CustomEventHandler
 {
-  typedef std::unordered_map<AL::event::EventId, uint32_t> EventToMaya;
+  typedef std::unordered_map<AL::event::EventId, size_t> EventToMaya;
 public:
 
   /// \brief  a structure to store some binding info for the MMessage event info
